requirements.txt
langchain
wikipedia
langchain-community
langchain-google-genai
langchain-openai
langchain-anthropic
python-dotenv
pydantic
PyGreSQL
psycopg2
duckduckgo-search